Fix documentation generation
This was working by coincidence rather than as intended. The Makefile
step to generate .rst files for each plugin was not working. However,
because in this module the plugins are in the toplevel package, the
call to sphinx-apidoc was generating .rst files for them instead.

The problem with that approach is that the plugin docs got a bunch
of useless function references at the bottom, where we should only
be including the docstring at the top.

To fix this, the sphinx-apidoc call is removed (this module doesn't
have any public API other than the plugins themselves) and the
Makefile step to generate .rst files is copied more closely from the
one in buildstream.git.