x86image plugin: Avoid moving the whole filesystem around needlessly
On the default Linux platform, BuildStream creates the staging area using
hardlinks that point back into the cache. If we `mv` the entire contents of
the staging area into /buildstream then we end up duplicating all of that
data rather than just moving the hardlinks, so it is very slow.

This patch changes the plugin to build an ext4 filesystem directly from the
%{build-root} directory which ensures we copy the data directly into the
disk image.

This reduces a build of a 10GB VM image from 50 minutes to 25 minutes for me.