docker source: Handle exception when the server doesn't return us JSON
This is a frequent occurence when we are given an incorrect URL for the
image. If the server returns an error code then we get an exception from
'requests', but if the server returns "200 OK" and a bunch of HTML then
we get this exception.

We do set the Accept header in the request to the specific MIME types
we want but we do still get HTML back in many cases.