sources/docker.py: Fix get_ref() to work when digest is not set
Currently, `get_ref()` does not work when `self.digest` is not set as it
assumes that it is a string, which is not the case. This means that
tracking a docker source which does not has any ref set will fail with
a no-so-friendly message saying
`AttributeError: 'NoneType' object has no attribute 'startswith'`.

Fix it so that `get_ref()` returns `None` when the ref is `None`.