sources/docker: Separate registry-url and image name keys
Currently `docker` source plugin accepts only a single `url` parameter
for the full image path, including the registry endpoint. And later it
splits them assuming that hostname part of the url is the registry
endoint and the rest is the image path. This causes issues when the
endpoint is in fact located at a subpath on that hostname, for example
'http://example.com/my-docker-registry'.

Fix it by explicitly asking for `registry-url` and `image` separately.

As this is a breaking change, bump the version number. Note that the old
format is now unsupported.

Fixes #9.