Add git-tag plugin
In freedesktop-sdk we have wanted a way to track the latest tag on
a git branch for a long time. This [1] BuildStream branch implemented
the changes necessary, but it looks unlikely to ever be merged, see
the discussion here [2]. As a result I have added the functionality
into a plugin we can use instead of the BuildStream git plugin.

I had to use importlib to access the git plugin, but doing this
makes this plugin far easier to maintain, as only 3 methods will
actually require maintenance: GitMirror.latest_commit(),
GitSource.configure() and GitSource.track(), the rest are inherited
from the BuildStream plugin.

[1] buildstream!303
[2] buildstream#220