git_tag: Add capability to track two branches
In freedesktop-sdk we want to track the latest stable tag, however this
may well be on master when there's a new stable release. As a result it
is desirable to track two branches, and track to the latest commit in
either, for human review of whether the updated ref is suitable. The
idea of the workflow this enables is:
* Automated job tracks elements and finds new tag
* Automated job checks new tag ABI compatible
* MR automatically opened to update the element
* Human reviews whether the bot-proposed new tag is stable
* MR is accepted/denied based upon review, human can now edit tracking
params on element if new major release branch