Force updating tags when fetching git repository
Backport of BuildStream/buildstream!1000.

When using aliases there are multiple remotes used in the cache
repository. When fetching it does not allow to updating tags from a
different remote.

So if you have the combination of changing upstream, renaming aliases,
adding mirrors in the same time as a tag having been changed, we end
up not fetching tags.

Commits that not in a branch and only tagged do not get fetched if the
tag is not fetched.