git_tag.py: Enable glob matching on returned tags
It may be possible to reduce manual checking of the suitability of
automatically returned tags via glob matching, for example if one only
wants stable releases to be tracked, it would be appropriate to exclude
tags denoting a release candidate, like v*-rc*.

This uses the `--match` and `--exclude` arguments to allow users of the
plugin to specify globs which can be excluded or matched for when
tracking to the latest tag.

I also fix a bug caused by a missing `else`, which actually stopped the
tag from being used when updating.