git_tag: Allow tracking multiple branches
Allow configuration of a list of additional branches to track, running
`bst track` will get the latest valid commit on any of the branches.
In conjunction with the `match` and `exclude` attributes this should
reasonably allow a user to automatically find the latest stable release
of a source without too many false positives.