Add test for x86image plugin
Adds a test for the x86image plugin. This uses the minimal image from
freedesktop-sdk, as I think this should be faster than building a
kernel, bootloader and the necessary dependencies from the alpine image.

Also uses some tooling from freedesktop-sdk in order to run the tests.

This will probably add a non-trivial amount of time to the test suite,
largely just to pull the dependencies we need for the image, but this is
simpler than adding a load of build dependencies for the image.