Move cmake plugin to bst-plugins-experimental
It was agreed on the mailing list to move all plugins into a single
repository, before moving them into domain specific repositories. As a
result it seems reasonable to move everything to the
bst-plugins-experimental repo as this stepping stone, rather than
creating a whole new repository.

This commit starts the process of moving things over by moving only the
cmake plugin along with its tests.