ostree: Disable GVfs to avoid background threads in the main process
If the DBus-based GVfs backend is installed, creating an OSTree Repo
object using Gio.File will spawn background threads.

As Source.get_consistency() is called in the main process, this can
cause problems with BuildStream's fork-based process model.

This simply disables the DBus-based GVfs backend as it's anyway not
needed in this context.