Add bazel plugins
Adds a bazel BuildElement and a bazel Source, which should really be
used together. The design decisions should be reasonably well documented
in the plugins themselves, but as an overview the workflow is this:

1. Create a `kind: bazel` element.
2. Add a source to a bazel project
3. If the bazel project does not have a repository resolved file, then
   add one as a local/patch source.
4. Add a bazel source.
5. The bazel source will download all of the external bazel dependencies
   specified in the repository resolved file
6. The downloaded sources are passed to the `bazel build` command using
   the `--distdir` option.
7. Bazel successfully builds your target inside a sandbox.

The repository resolved file is generated using `bazel cquery
--experimental_repository_resolved_file=string deps(//foo:bar)`
This is compatible with every documented version of bazel.

The source plugin may be configured to fall back on host bazel to
generate a repository resolved file, however this loses the ability to
have the ref verified, as the repository resolved file is not generated
deterministically. Best practice would be not to use this.

The bazel build element requires /dev/shm to be in the sandbox, so
depends on this MR [0] for BuildStream. We also make sure that /home
exists, and call it $HOME. This is because bazel makes a heavy usage of
$HOME/.cache, and we want this separate from both build and install root
to make sure we don't contaminate the finished binaries.

Finally, at the moment this doesn't support remote execution or caching
for bazel, which largely removes the benefits. Downloading the remote
cache could potentially be done during the source fetch, and pushing in
a separate stage of commands after the build is completed, in the
sandbox but with network access.

[0]: https://gitlab.com/buildstream/buildstream/merge_requests/1694