setup.py: Remove old-style buildstream.plugins entrypoints
Since BuildStream 1.93.0, the entry points for source and element
plugins are separate. Previously, bst-plugins-experimental was modified
in a backwards compatible to ensure that BuildStream tests keep passing.

Now that we have a new BuildStream release, we can drop the old style
entry point definition. Note that this drops support for
BuildStream version < 1.93.0.