tests/sources/deb.py: Fix tests
This seems to have been an error during the migration of these tests
over from the main BuildStream repository. These tests were incorrectly
using the core `deb` plugin instead of the plugin in this repository.

This wasn't caught before as the pinned version of BuildStream in this
repository still contained the `deb` source. But that is removed in
BuildStream 1.93.0.

Similarly, `HAVE_ARPY` is no longer defined in BuildStream. Although in
this case, we can actually drop the check since we always the tests with
the plugin dependencies installed