.gitlab-ci.yml: Build documentation on each commit
Although we wish to deploy documentation only on pushes to master
branch, we should still verify that it builds correctly with each
commit.

So, separate the documentation job into two parts - one that builds it
and one that publishes it. The latter is only run on master.

The documentation job was also doing more manual steps to install
dependencies, but is unnecessary now that we have tox environment for
the same.