sources/cargo.py: Fix alias
URLs need to be "marked". The plugin marked the base URL instead of
the URLs of the crates. This broke an assertion:

```
File "/usr/lib/python3.7/site-packages/buildstream/source.py", line 681, in mark_download_url
    ), "URL was not seen at configure time: {}".format(url)
AssertionError: URL was not seen at configure time: crates:crates
```

This is a port of BuildStream/bst-external!115