bazelize: Add support for cc_import rules
`cc_import` is intended for including pre-compiled libraries. Any entry
satisfying the following will be treated as `cc_import`:

* does not have compilation unit sources
* does not have dependencies, link options, or compilation options
* has only one source file matching one of the appropriate library
regexes

There is some overlap in the allowed extensions of shared, static, and
interface libraries:
* "*.lib" overlaps in static and interface (assumed to be
static)
* "*.so" overlaps in shared and interface (assumed to be shared)
* "*.dylib" overlaps in interface and shared (assumed to be shared)

It is not possible to decide at a bst level if linking will succeed
with the values passed so these values are collected based on matching
the extensions of the files.