bazelize.py: Adapt to removed Scope type
This plugin was illegally using Scope.ALL, but this is irrelevant,
it only means it was picking up dependencies which would not be
present for a build.

This plugin was also trying to bypass BuildStream API paths to access
API directly on dependency Bazel elements, which is illegal and now
impossible.

This plugin could be simplified by staging the dependency bazel elements
and collecting the bazel BUILD files, and then concatenating them in
the bazelize element's assemble() method, but for now we're just taking
a simpler approach and using Element.set_public_data() and
Element.get_public_data() to fix the plugin without a major refactor.